Extract CORTAS operation on delta table to a dedicated test
